penumbra_sdk_auction::auction
Module dutch

Re-exports

pub use actions::ActionDutchAuctionEnd;

pub use actions::ActionDutchAuctionSchedule;

pub use actions::ActionDutchAuctionWithdraw;

Modules

actions

Structs

DutchAuction

A deployed Dutch Auction, containing an immutable description and stateful data about its current state.

DutchAuctionDescription

A description of the immutable parts of a dutch auction.

DutchAuctionState

A stateful description of a dutch auction, recording its state (via a sequence number), the current position id associated to it (if any), and its amount IO.

Constants

DUTCH_AUCTION_DOMAIN_SEP