penumbra_sdk_auction
Module event

Functions

auction_vcb_credit

auction_vcb_debit

dutch_auction_closed_by_user

Event for a Dutch auction that is ending because it has been closed by its owner.

dutch_auction_exhausted

Event for a Dutch auction that is ending because it has been completely filled.

dutch_auction_expired

Event for a Dutch auction that is ending because it has expired.

dutch_auction_schedule_event

Event for a Dutch auction that has been scheduled.

dutch_auction_updated

Event for an execution round of a Dutch auction.

dutch_auction_withdrawn

Event for a Dutch auction that is withdrawn by a user after ending.